+++ src/js/PlotBoilerplate.js
@@ -54,7 +54,8 @@
  * @modified 2020-03-28 Ported this class from vanilla-JS to Typescript.
  * @modified 2020-03-29 Fixed the enableSVGExport flag (read enableEport before).
  * @modified 2020-05-09 Included the Cirlcle class.
- * @version  1.7.3
+ * @modified 2020-06-22 Added the rasterScaleX and rasterScaleY config params.
+ * @version  1.8.0
  *
  * @file PlotBoilerplate
  * @fileoverview The main class.
@@ -157,6 +158,8 @@ var PlotBoilerplate = /** @class */ (function () {
             offsetX: PlotBoilerplate.utils.fetch.num(config, 'offsetX', 0.0),
             offsetY: PlotBoilerplate.utils.fetch.num(config, 'offsetY', 0.0),
             rasterGrid: PlotBoilerplate.utils.fetch.bool(config, 'rasterGrid', true),
+            rasterScaleX: PlotBoilerplate.utils.fetch.num(config, 'rasterScaleX', 1.0),
+            rasterScaleY: PlotBoilerplate.utils.fetch.num(config, 'rasterScaleY', 1.0),
             rasterAdjustFactor: PlotBoilerplate.utils.fetch.num(config, 'rasterAdjustdFactror', 2.0),
             drawOrigin: PlotBoilerplate.utils.fetch.bool(config, 'drawOrigin', false),
             autoAdjustOffset: PlotBoilerplate.utils.fetch.val(config, 'autoAdjustOffset', true),
@@ -537,8 +540,8 @@ var PlotBoilerplate = /** @class */ (function () {
      * @return {void}
      **/
     PlotBoilerplate.prototype.drawGrid = function () {
-        var gScale = { x: Grid_1.Grid.utils.mapRasterScale(this.config.rasterAdjustFactor, this.draw.scale.x),
-            y: Grid_1.Grid.utils.mapRasterScale(this.config.rasterAdjustFactor, this.draw.scale.y) };
+        var gScale = { x: Grid_1.Grid.utils.mapRasterScale(this.config.rasterAdjustFactor, this.draw.scale.x) * this.config.rasterScaleX,
+            y: Grid_1.Grid.utils.mapRasterScale(this.config.rasterAdjustFactor, this.draw.scale.y) * this.config.rasterScaleY };
         var gSize = { width: this.grid.size.x * gScale.x, height: this.grid.size.y * gScale.y };
         var cs = { width: this.canvasSize.width / 2, height: this.canvasSize.height / 2 };
         var offset = this.draw.offset.clone().inv();
@@ -772,7 +775,7 @@ var PlotBoilerplate = /** @class */ (function () {
      * @return {void}
      **/
     PlotBoilerplate.prototype.clear = function () {
-        // Note that the image might have an alpha channel. Clear the scene first.
+        // Note that elements might have an alpha channel. Clear the scene first.
         this.draw.clear(this.config.backgroundColor);
     };
     ;