+++ src/js/interfaces.js
@@ -1,4 +1,3 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-// export const DRAGGABLE_VERTEX:string = 'vertex';
 //# sourceMappingURL=interfaces.js.map